+++ modules/particles/src/SkParticleEffect.cpp
@@ -40,18 +40,29 @@ SkParticleEffect::SkParticleEffect(sk_sp<SkParticleEffectParams> params, const S
 }
 
 void SkParticleEffect::start(const SkAnimTimer& timer, bool looping) {
+    this->start(timer.secs(), looping);
+}
+
+void SkParticleEffect::start(double now, bool looping) {
     fCount = 0;
-    fLastTime = fSpawnTime = timer.secs();
+    fLastTime = fSpawnTime = now;
     fSpawnRemainder = 0.0f;
     fLooping = looping;
 }
 
 void SkParticleEffect::update(const SkAnimTimer& timer) {
-    if (!timer.isRunning() || !this->isAlive() || !fParams->fDrawable) {
+    if (!timer.isRunning()) {
         return;
     }
 
-    double now = timer.secs();
+    update(timer.secs());
+}
+
+void SkParticleEffect::update(double now) {
+    if (!this->isAlive() || !fParams->fDrawable) {
+        return;
+    }
+
     float deltaTime = static_cast<float>(now - fLastTime);
     if (deltaTime < 0.0f) {
         return;